Set ldflag -w to turn off DWARF debugging information
Stripping DWARF debugging information has some drawbacks:

- gdb can not be used on the binary to look at specific functions, set
  breakpoints or get stack traces, because all the metadata gdb needs
  will not be included.
- same on other tools depend on the information, like pprof profiling

LDFLAGS in config.mk has also the -s option that turns off generation of the
Go symbol table so 'go tool nm' can not be used to list the symbols in the binary.

If need to profile or debug the binary, this options must be disabled.